requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
DateTime==5.5
decorator==4.4.2
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.4.2
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
packaging==24.2
pillow==10.4.0
proglog==0.1.11
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
six==1.17.0
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
typing_extensions==4.13.1
urllib3==2.3.0
zope.interface==7.2
